+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+### Added
+
+- Adds the functionality from the cirrus api lambda to the `Deployment` class and
+  the CLI. ([#324])
+
 ### Fixed
 
 - The update state lambda was not correctly parsing the error and cause from the Step
@@ -21,7 +26,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   unchecked.  Moto mock testing did not complain that the State Machine ARN
   provided was invalid for StepFunctions execution. ([#317])
 - Dynamo DB fails to update with extraneous parameters in the
-  `ExpressionAttributeValues` argument.  Issue not surfaced by Moto. (#319)
+  `ExpressionAttributeValues` argument.  Issue not surfaced by Moto. ([#319])
 
 ## [v1.0.1] -- 2025-06-18
 
@@ -1104,6 +1109,7 @@ Initial release
 [#315]: https://github.com/cirrus-geo/cirrus-geo/pull/315
 [#317]: https://github.com/cirrus-geo/cirrus-geo/pull/317
 [#319]: https://github.com/cirrus-geo/cirrus-geo/pull/319
+[#324]: https://github.com/cirrus-geo/cirrus-geo/pull/324
 [#326]: https://github.com/cirrus-geo/cirrus-geo/pull/326
 [f25acd4]: https://github.com/cirrus-geo/cirrus-geo/commit/f25acd4f43e2d8e766ff8b2c3c5a54606b1746f2
 [85464f5]: https://github.com/cirrus-geo/cirrus-geo/commit/85464f5a7cb3ef82bc93f6f1314e98b4af6ff6c1